backend-centralized/requirements.txt
# Complete Requirements - Generated from working virtual environment
# This includes all successfully installed and tested packages

# Core Framework
fastapi==0.116.1
uvicorn[standard]==0.35.0
websockets==13.0
starlette==0.47.2
anyio==4.10.0
h11==0.16.0
click==8.1.7
sniffio==1.3.1

# Database
SQLAlchemy==2.0.42
alembic==1.16.4
psycopg2-binary==2.9.10
greenlet==3.2.4
Mako==1.3.10
MarkupSafe==3.0.2

# Authentication & Security
python-jose==3.5.0
passlib==1.7.4
python-dotenv==1.1.1
cryptography==45.0.6
PyJWT==2.8.0
bcrypt==4.1.2
ecdsa==0.19.1
rsa==4.9.1
six==1.17.0
cffi==1.17.1
pycparser==2.22

# Data Validation
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
email_validator==2.2.0
dnspython==2.7.0
annotated-types==0.7.0
typing-extensions==4.14.1
typing-inspection==0.4.1

# Background Tasks & Caching
redis==6.4.0
celery==5.5.3
billiard==4.2.1
kombu==5.5.4
vine==5.1.0
amqp==5.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
prompt_toolkit==3.0.51
wcwidth==0.2.13

# Payment Processing
stripe==7.6.0

# Communication Services
twilio==8.10.0
aiohttp-retry==2.9.1
aiohttp==3.12.15
requests==2.32.4
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
attrs==25.3.0
frozenlist==1.7.0
multidict==6.6.3
propcache==0.3.2
yarl==1.20.1
charset-normalizer==3.4.3
urllib3==2.5.0
certifi==2025.8.3
python-multipart==0.0.20

# Push Notifications & Firebase
firebase-admin==6.4.0
google-api-python-client==2.178.0
google-cloud-storage==3.2.0
google-cloud-firestore==2.21.0
google-auth-httplib2==0.2.0
google-resumable-media==2.7.2
google-crc32c==1.7.1

# Google Cloud Services & gRPC
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
protobuf==6.31.1
proto-plus==1.26.1
uritemplate==4.2.0
httplib2==0.22.0
pyparsing==3.2.3

# Caching and Authentication
CacheControl==0.14.3
cachetools==5.5.2
pyasn1-modules==0.4.2
pyasn1==0.6.1

# Utilities
python-dateutil==2.9.0.post0
packaging==25.0
tzdata==2025.2
idna==3.10
msgpack==1.1.1

# Monitoring & Logging
structlog==25.4.0
apscheduler==3.10.4

# Rate Limiting
slowapi==0.1.9

# Enhanced Real-time Features
websocket-client==1.8.0
python-socketio==5.11.1
python-engineio==4.9.1

# External Service Integrations
sendgrid==6.11.0
python-telegram-bot==20.8
slack-sdk==3.36.0
discord.py==2.5.2

# Enhanced Analytics & Monitoring
psutil==6.1.0
prometheus-client==0.20.0
datadog==0.48.0

# System Monitoring
psutil==6.1.0
GPUtil==1.4.0

# Enhanced Security
python-multipart==0.0.20
itsdangerous==2.2.0
Werkzeug==3.1.3

# File Processing & Storage
Pillow==11.1.0
python-magic==0.4.27

# Enhanced Communication Features
asyncio-mqtt==0.16.1
paho-mqtt==1.6.1

# Database Migrations & Management
alembic==1.16.4
sqlalchemy-utils==0.41.1

# Enhanced Error Handling & Logging
loguru==0.7.2
sentry-sdk==2.17.0

# Performance & Optimization
# uvloop==0.20.0  # Not compatible with Python 3.13
orjson==3.11.2

# Testing & Development
pytest==8.2.2
pytest-asyncio==0.24.0
httpx==0.28.1

# Note: Video processing packages (opencv, av, etc.) are excluded
# as they can cause build issues on some systems. Install separately if needed.
# For full backend camera features, uncomment:
opencv-python-headless==4.10.0.84
av==12.1.0